# Adding request start time on Linux.