--- a/include/haywire.h
+++ b/include/haywire.h
@@ -6,6 +6,7 @@ extern "C" {
 
 #include <stdio.h>
 #include <stdbool.h>
+#include <time.h>
 
 #ifdef _WIN32
     /* Windows - set up dll import/export decorators. */
@@ -161,6 +162,9 @@ typedef struct
     void* headers;
     hw_string* body;
     int body_length;
+#ifdef __linux__
+    struct timespec* start_time;
+#endif
 } http_request;
 
 typedef	void* hw_http_response;