--- a/src/haywire/http_request.c
+++ b/src/haywire/http_request.c
@@ -101,6 +101,10 @@ http_request* create_http_request(http_connection* connection)
     request->body = malloc(sizeof(hw_string));
     request->body->value = NULL;
     request->body->length = 0;
+#ifdef __linux__
+    request->start_time = malloc(sizeof(struct timespec));
+    clock_gettime(CLOCK_REALTIME, request->start_time);
+#endif
     INCREMENT_STAT(stat_requests_created_total);
     return request;
 }
@@ -123,6 +127,9 @@ void free_http_request(http_request* request)
     }
     free(request->url); 
     free(request->body);
+#ifdef __linux__
+    free(request->start_time);
+#endif
     free(request);
     INCREMENT_STAT(stat_requests_destroyed_total);
 }