This has no significant performance impact as far as I can see (after switching to gettimeofday):
### master

```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 50 -t 50 -d 10s -s pipelined_get.lua  --latency http://172.31.5.65:8000 -- 64
Running 10s test @ http://172.31.5.65:8000
  50 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.42ms    1.50ms  40.03ms   60.71%
    Req/Sec    14.47k     4.58k   53.01k    77.62%
  Latency Distribution
     50%    2.47ms
     75%    3.96ms
     90%    5.54ms
     99%    0.00us
  7266329 requests in 10.09s, 1.19GB read
Requests/sec: 720072.32
Transfer/sec:    120.86MB
```
### this branch

```
Running 10s test @ http://172.31.5.65:8000
  50 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.51ms    1.55ms  25.95ms   57.70%
    Req/Sec    14.47k     5.68k   54.31k    81.64%
  Latency Distribution
     50%    2.59ms
     75%    4.26ms
     90%    5.65ms
     99%    0.00us
  7266703 requests in 10.09s, 1.19GB read
Requests/sec: 719979.96
Transfer/sec:    120.85MB
```