@nmdguerreiro 2016-01-27
Actually, the latency increases substantially (I was only looking at the throughput :cry: ). Closing the PR