@nmdguerreiro 2016-01-27
After switching to gettimeofday and removing an extra allocation, results look really good:

### this branch
```

Running 10s test @ http://172.31.5.65:8000
  50 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.51ms    1.55ms  25.95ms   57.70%
    Req/Sec    14.47k     5.68k   54.31k    81.64%
  Latency Distribution
     50%    2.59ms
     75%    4.26ms
     90%    5.65ms
     99%    0.00us
  7266703 requests in 10.09s, 1.19GB read
Requests/sec: 719979.96
Transfer/sec:    120.85MB
```