@rzezeski 2016-01-28
> This has no significant performance impact as far as I can see (after switching to gettimeofday):

This is because of vDSO. It avoids the switch into kernel mode by mapping the clock value into userspace. OSX does a similar thing, but calls it the "comm page". In an illumos-based OS (e.g SmartOS) it will take a "fast trap", it's a lightweight syscall trap but still takes about an order of magnitude more than vDSO or comm page. But we are still talking only hundreds of nanoseconds on a modern Intel processor. I'm not sure if BSD folks play similar tricks. I know FreeBSD has discussed adding something like comm page, but not sure if it ever made it in.

Anyways, just wanted to shed some light on why `gettimeofday(2)` doesn't add much overhead on most operating systems. However, if Haywire ever dips into microsecond response times then perhaps something like this could count as overhead.