@rzezeski 2016-01-28
Also, is there a reason the 99% reports 0.00? Makes me wonder if there is something wrong with the benchmark client.