@kellabyte 2016-01-29
@rzezeski I don't know the reason but I think `wrk` is messed up somewhere. But even before that bug existed, `wrk` was known not to compensate for coordinated omission so you can't really trust its latency distribution report anyways. I use [`wrk2`](https://github.com/giltene/wrk2) from @giltene when I want a proper latency report.

@rzezeski We do get into sub-millisecond response times very easily if we aren't slamming the box. On a machine that does 9.2M req/s if we give it say a 1M req/s or 500,000 req/s workload most of the requests are under 1ms.

If we start instrumenting begin/end request timestamps that's going to be `requests/second * 2`. Is this going to interfere with response times?