@rzezeski 2016-01-29
> If we start instrumenting begin/end request timestamps that's going to be requests/second * 2. Is this going to interfere with response times?

On Linux, on a fairly modern Xeon, I've seen gettimeofday run about 50ns. Let's round that up to 100ns to be safe. That's 200ns per request. Or about 0.02% for a 1ms response time. I'd say if you dip below 10μs then you start to worry.

That said, gettimeofday may not be cheap on all systems. It may be more like 500-1000ns on a different platform. In that case it would be 0.2% of a 1ms response time. That's still too small to matter but if you get into tens of microseconds then it might be worth removing.

And one more thing, if you simply want to get a time delta then there is a better call to use: `clock_gettime()` with `CLOCK_MONOTONIC`. That will give you nanoseconds since some arbitrary point in time. This clock may be more efficient to query on some systems since it ignores the system clock and typically uses something like the TSC (Time Stamp Counter) found on Intel.