@rzezeski 2016-01-29
I forgot to mention, `clock_gettime()` is defined in POSIX. So it should have wide support.

http://pubs.opengroup.org/onlinepubs/9699919799/functions/clock_getres.html