@nmdguerreiro 2016-01-30
@rzezeski I did try to use clock_getttime initially, but the resolution on my servers was 10ms, which is far too coarse for my use case.