@kellabyte 2016-01-31
I'm not sure I like adding this to the code base that requires every single request to include.

Currently Haywire is doing `9.2M requests/second` with `pico` parser with a `32` batch size in an experimental branch and we are doing `1.4GB/second` on a network that is capable of `2.2GB/second` so theoretically we could be going something like `16M requests/second`.

A `100ns` call to get time every request can go a maximum of `10M/second` and will use 100% CPU usage and that's excluding any Haywire code fitting in. I think this will impact performance quite a bit and handicap further optimizations.

I think it's better if we offer a `khash` dictionary on the request where application code/haywire code can add statistics and we provide some handy callbacks like `before_request` and `after_request` the application can implement and populate request statistics if this kind of performance isn't an issue.

Thoughts?